Fixes some remaining report-generation issues and increases generation speed.
This commit updates the ColorBoxPlot class to accept the new objective
function builder objects in addition to simple 'logl' or 'chi2' plot
types.  It also removes the "directlogl" and "directchi2" options from
the available plot types, as these were experimental and aren't used
anymore.  A fair amount of code in plothelpers.py has been commented
and marked w/REMOVE that should be unnecessary under the new objective
function framework.  Caching is still not fully used, and the CACHE
placeholder persists for later enabling it.

Minor updates to the report-notebook code was made so that a fully
runnable report notebook is now generated again.

An inquiry into why it takes so long to generate even a small 1Q
report led to the realization that lots of time is spent within
plotly validation functions, as well as the new "gauge robust" tables
generated in section/gauge.py.  This commit adds "BlankTable" comments
above these slow tables (just to mark them), and does the following
things to speed up the plotly validation:
1) sets a default plotly theme instead of calling update_layout
   on every created figure, and sets validation=False so that
   validation at this stage is turned off (see plotly_plot_ex.py
   and VALIDATE_PLOTLY in particular)
2) in matrix_color_boxplot (in workspaceplots.py) the returned
   plotly go.Figure object and its contained go.Layout have been replaced
   with normal dicts.  This was given on plotly forums as a workaround
   for the slow validation that takes place when using the graph objects
   (hopefully something nicer in the future?).  Validation at the time
   of figure creation is still a bottleneck in the 1Q case, so we could
   continue this replacement across more of pygsti - but it would be
   nice to have a clean way of marking where we've done this that would
   allow us to switch back to the graphy objects if desired.
3) adds tol=1e-5 and maxiter=100 arguments to the gauge optimizations
   performed to create the gauge-robust tables.  This speeds up the
   table generation a little, but ideally we'd find a better way to
   find the "best gauge" for a given gate.