Begins implementation of create_from paradigm.  Fixes test_core units.
This commit updates the "build" classmethod in a number of classes
to "create_from" since we want to reserve using the name "build"
for builder-paradigm situations.

This commit also updates a number of minor issues with the objective
functions and updates test_core.py so that the new core.py functionality
(and objectivefns.py) are being tested by the unit tests again.  All
tests in test_core.py pass now.