Adds TypedDict and updates NamedDict to improve conversion to DataFrames.
The TypedDict class is meant to allow users to specify an entire
set of values to place at the leaves of a nested NamedDict tree
(though it's not currently used).  This commit updates NamedDict
and the nested-NamedDict creation functions in protocol.py to
make the conversion to data frames more concise.  It also begins
removing spaces from DataFrame column names (convenient for
accessing columns as attributes), as it seems good to establish
a CamelCase convention for column names.