Updates ProtocolResultsDir as_dataframe to have a Path column.
This update changes the way a ProtocolResultsDir dataframe is
constructed.  Previously the dataframe would have a column for each
layer of path hierarchy, and these layers would (necessarily) be
named different things like EdesignBranch and Qubits.  Now, the
path to a given node is stored in a single "Path" column, and there
are no other columns tied to the path within the edesign tree.
Because "Qubits" is a useful column, this commit adds a "Qubit"
column via ExperimentDesign, which adds self.qubit_labels as
to its named dict via _nameddict_attributes.