Updates naming in VB and RB protocols, improves results->dataframe conversion.
Renames the SummaryStatsConstructor protocol to SummaryStatistics, and
updates "datatype" to "statistic" throughout. SummaryStatistics acts
as a base class for protocols like RB or VB variants.  Updates to the now-named
SummaryStatisticsResults allow it to hold arbitrary information
and render it nicely to a dataframe (by overriding
ProtocolResults._my_attributes_as_nameddict).

The ByDepthSummaryStatistics protocol has been updated to be able to compute
multiple (instead of just one) statistic, utilizing the updates to
the results object.