Fixes a number of failing system tests.  Updates memory limiting slightly.
Updates several failing system tests that were outdated, and updates
two parts of the memory limit framework in order to make memory limits
more robust:
1) sets the "gather-memory" limit (the memory used in cross-processor
gather operations) based on the amount of available memory rather than
the just the limit alone.  This accounts for the "current memory", which
has shown itself difficult to obtain from the system, but this seems
better than setting the gather-limit based just off of the limit alone.

2) Changes the trigger for memory errors to include the current memory,
i.e. we raise an error if (mem_limit - current_mem) < needed_mem rather
than just if mem_limit < needed_mem.  This seems more consistent with
the rest of the memory calculations (but as mentioned above, relies on
an accurate current-memory estimate which may be problematic).