Adds "frobeniustt" metric to gaugeopt_to_target(...).
The `gates_metric` and `spam_metric` arguments can now be set to
"frobeniustt".  This computes the frobenius distance between a
*transformed-target* model and the (non-transformed) initial model.
This is an experimental alternative that may become the default, as
the frobenius distance between a gauge-transformed model and a target
is *not* symmetric with respect to which model is transformed, and
it may be better behaved when the target is transformed.

This can be seen, for instance, when the target has a perfect identity gate:
the gauge optimization really shouldn't care how the identity is
transformed, but it *will* if we transform a near-identity estimate.