Integrates target-transform gauge optimization - adds "stdgaugeopt-tt".
This commit fixes some bugs and sets up standard pyGSTi workflows
with the option to use gauge optimizations that minimize the frobenius
distance between an estimated model and the gauge-transformed *target* model.
This "frobeniustt" metric (as specified to `gaugeopt_to_target`) is
used in the newly available "stdgaugeopt-tt" suite name, and is now used
throughout the gauge-robust gate metrics table.

This functionality has been lightly tested and appears to be working.