Fixes an issue with setting the default plotly template.
We've changed the way we set the "template" used for our plotly
plots several times prior to this.  Setting the template directly
via a "fig.update_layout(template=...)" call on every figure doesn't
work well because plotly validates the entire figure as a part of
this call, making it very *slow* for our large (boxplot) figures.

We sidestepped this issue by instead setting a default template.
However, we did so by setting it within a function instead of
at the top of the plotly_plot_ex.py module, which meant it
(sometimes?) didn't work.  The default plotly template must be
set *before* any plotly figures are constructed, and the plotly_plot_ex
functions are called at *render*-time, which is after the figures
are created.

This commit fixes this issue by moving the default-template-setting
to the beginning of the module.  It guards the import in a try/except
block because setting the template isn't an essential operation, and
even though pyGSTi technically requires plotly, it can perform many
things without it.