Fixes a bug in the display of color boxplots.
Updating the color box plots to use the new objective function
objects introduced a bug causing many/all of the plaquettes in a
boxplot to look the same.  This was caused because a plaquette
performed a local "circuit simplification" and expected
`elementvec_to_matrix` to be passed a *local* array of the values
just for that plaquette.  Instead, it was passed the global
terms of the objective function (the same given to all plaquettes),
causing the duplication.

This commit fixes this issue *and* helps to simplify the circuit
structure objects (anticipating more work on this in the near
future) by removing the local circuit simplification from a plaquette.
The `iter_simplified` and `simplify_circuits` methods have been
removed from a CircuitPlaquette, and similarly `simplfy_plaquettes`
has been dropped from CircuitStructure.  Now the `elementvec_to_matrix`
function receives a ready-built dictionary telling it where to find
the values corresponding to its (and potentially other) circuits,
and there's no need for the plaquette to store element-index arrays
locally.