Updates "operation sequence" => "circuit" where appropriate in docstrings.
This replacement is performed across the entire repository.  At some
earlier point, we were thinking of using "operation sequence" instead
of circuit, but we've commited to using "circuit" now.  We still use
"operation sequence" more or less synonymously, but to mean in particular
a circuit without any prep or measurement (no SPAM).