Adds circuit-probability-array objects (to replace evaluation trees).
The CircuitOutcomeProbabilityArrayLayout class defines a mapping
between array indices and the outcomes for all the circuits in a
list.  This object will associate the "element indices" or "lookup"
dictionaries with related information (like the circuit list) to
better organize this informaiton.  "Evaluation strategy" classes
subclass CircuitOutcomeProbabilityArrayLayout and provide, in
addition to the layout (which is based on resources like the # of
processors), additional structure to support the fast compuation
of array elements using a *specific* forward simulator.  Generally,
a forward simulator will be told to compute something in an array
whose structure is given by a layout (or strategy) object.

This reorganization will allow strategy objects to create sub-strategies
(e.g. sub-trees) at the *same* time as they are setting up the
COP array layout, removing the need to perform a ".split(...)" after
many complex internal bookeeping quantities are formed.  This will
dramatically simplify the evaltree => evalstrategy objects, as much
of their complication centered around split operations.

A WIP - this is just the first step; nothing is working yet.