Adds more meat to evaluation strategy classes.
Now evaluation strategies have "atoms" - objects which encapsulate
an atomic division of work and hold everything needed for (a single
processor) to apply the strategy to a portion of the entire
circuit outcome probability array (COPA).

Lots of functions are commented out in these file, which we'll deal
with as we figure out which ones are really needed and which should
be a part of the "Atom" classes.