Updates COPA layout classes to merge "eval tree" functionality with layouts.
This commit renames what were "evaluation tree" objects to "layout"
objects, as they are now sub-classes of COPALayout - the object that
maps circuit outcomes to array indices.

The "evaluation tree" logic has been largely simplified because now
you create an evaluation tree/strategy at the *same time* a layout
is decided upon.  This means that the horrific "split" functionality,
which needed to essentially split an *existing* layout into pieces,
is no longer needed.  A much simpler approach whereby index partitioning
is decided upon before computing various layout quantities (used
to speed up forward simulations) are computed.

The new DistributableCOPALayout class replaces the "EvalTree" base
class, and introduces a framework whereby subtrees are called "atoms"
(more generically), and these are the pieces allocated to processors
to perform computational work.  Currently, the paradigm is that each
processor will contiain a *full* copy of the final array, and partitioning
is used to reduce the size needed for intermediates (the "cache").  In
the future we may want to have a "gather=False" mode where each processor
only allocates a portion of the entire final array (FUTURE).