Updates forward simulator classes to use layouts.
This is only partially done.  This commit updates forward simulator
classes to work with layouts instead of evaluation trees.  It also
consolidates more of the forward simulator logic into the
DistributableForwardSimulator class, which works with a distributable
layout.  This commit also breaks up the "single-circuit" computational
routines into "simple" forward simulators that rely nearly entirely
on the base-class framework for implementing the required functionality.

This commit only updates the "matrix" and "map" simulators, and
only partially the "map" simulator because the replib functions still
need to be updated.  "term" and "opless" simulators will need to be
updated in the FUTURE.