Adds ModelDatasetCircuitsStore object and functionality.
Factors the "storage" part of what were MDSObjectiveFunction objects
(now MDCObjectiveFunction) into a new class: EvaluatedModelDatasetCircuitsStore,
derived from ModelDatasetCircuitsStore.  These classes implement a data
storage object for a given (model, dataset, circuit-list) tuple, and replace
the ComputationCache class.  Objective functions are now also "stores" that
hold information, so that it's trivial to pass on the cached values of a
GST run, for instance.