Applies fixes to get "map" fwdsim starting to work.
Fixes/tweaks the layout and forward simulator classes so they
actually run without errors.  The main updates of this commit are
to fastreplib.pyx, where we've updated the probability-computing
routines for the "map" forward sim.  We still need to update the
derivative & hessian routines, as well as slowreplib.py (though some
changes to slowreplib.py have been made as similar changes were
made to fastreplib.pyx).