Upgrades dirty-flag framework: nodirty => dirty_value arguments.
All ModelMember.from_vector functions now take a dirty_value argument
that tells it how to set the dirty flag after the call is complete.
The idea is that this flag will always be set to a definite value,
since either the from_vector call is isolated (so the dirty flag should
be set) or it's a part of the parent model's from_vector or similar
method, in which case the flag should be cleared (because the parent
model is syncing up its members).  The dirty_value argument replaces
the nodirty argument and makes the system clearer and more explicit,
both good things.  It also removes the need to recursively reset the
dirty flags in OpModel._cleanvec(...), as from_vector can now to
this via dirty_value=False.

This commit also adds the 'close' argument to Model.from_vector(...).