Updates replib: poly => polynomial renaming and term-fwdsim upgrade.
Upgrades term forward simulation routines so that now GST can be
run (1Q test example) using a pruned-path term simulator.  Updates
have been done on the fast and slow versions of the replib.

Also, this commit performs the poly => polynomial renaming throughout
the replib to be consistent with the rest of pygsti.