Upgrades time dependent calculations to use new fwdsim framework.
Gets the time-dependent (map-fwdsim) calculations working again.
This adds circuit information to the "atoms" of the MapCOPALayout
so that corresponding data can be accessed by the time-dependent
chi2 or logl calculation routines in replib (also updated here).

This commit improves the calculations so that TD-GST should now
work with instruments, though this will be inefficient because
the loops in the replib function loop over "expanded circuits"
and process a dataset row for each expanded circuit, even when
multiple expanded circuits will correspond to the *same* dataset row
when there are instruments.  This is an area for future upgrades.

"fast" and "slow" versions of the replib are updated, and the
TD-GST tutorial runs, though it should be tested more fully as the
result of TD-GST isn't as accurate as I would have expected.