Small updates to make the time-dependent calculations run faster.
Avoids repeated calls to DataSetRow.outcomes, which was a bottleneck.

Profiling the time-dependent calculations (with C-extensions
enabled) revealed that the majority of a TD-GST calculation is occupied
by the block:

for gl in remainder:
    op = opcache[gl] if (gl in opcache) else model.circuit_layer_operator(gl, 'op')
    op.set_time(t); t += gl.time  # time in gate label == gate duration?
    rho = op._rep.acton(rho)

within fastreplib.DM_mapfill_TDterms(...).  The time is roughly split
between the set_time() line and the acton(rho) line.  The former may
be sped up by identifying time-independent operations ahead of time
and not calling set_time on them.  The latter seems to be hard cost,
as acton is the raw action opertion in C-code.