Adds Circuit._fastinit method for faster circuit operations.
Many circuit methods return a new Circuit object, but Circuit.__init__
can be slow because it massages data passed to it, performing some
tasks automatically if needed.  A new _fastinit classmethod is now
used internally to more rapidly create a Circuit.  Care must be
taken to ensure, however, that the "editable" argument matches whether
the "layers" argument is either a tuple of list - this could be a
pitfall.