Cleans up model object modules.
Removes old/debug code that is no longer needed.  In particular,
the removal of all "fwdsim wrapper" functions except probabilities
and bulk_probabilities simplifies the model classes and reduces a
lot of long repeated docstrings.

This commit also removes the "dimension()" functions in Models and
ModelMember objects, preferring the shorter "dim" property, and
eliminating a 2nd way to do the same thing, which is bad practice and
can lead to confusion.