Fixes a number of cloud-noise and term-calc related bugs (from fwdsim upgrade).
This commit updates names and upgrades the usage of term forward simulators
within the cloudnoise-circuit selection, largely in nqnoiseconstruction.py.
One significant changes is that `polynomial_vindices_per_int` is no longer
an attribute of TermForwardSimulator (it was too hard to keep in sync with
the sim's .model) and instead this value is passed as an argument to many
of the term-calc related replib calculation functions.

Building cloud-noise models works now, and ciruit selection almost
works.  We still need to deal with the layerop_cache issues - that'll
come in another commit soon.