Fixes fwdsim layout atoms so they contain *local* not *global* element indices.
Certain members of a layout atom are meant to hold "global" element indices,
that is, the indices within the final array represented by the atom's parent
layout (e.g. atom.element_slice and the elindex_outcome_tuples dictionary
that is filled in during atom creation to implement the parent's .indices and
.outcomes methods).  However, members of the layout atoms that are used for
(local) computations should contain *local* indices, as the forward simulators
are filling local arrays (see DistributableForwardSimulator and how it passes
array_to_fill[atom.element_slice] to *_block methods).

This commit fixes the previous implementation which placed global indices
everywhere.  (The errors this fixes are only visible in a multi-processor
context, where local and global indices are different.)