Fixes an outstanding parsing issue with the time-dependent data text format.
The recently added "long" dataset text-file format (that allows specification
of time-resolved data) created an ambiguity that was not handled by
the parser.  Namely, when a circuit was on a line by itself it could have
been the start of a time-resolved "block" or just a circuit without any
corresponding data (following the old format).  This commit adds logic
to the parser that interprets a "blank circuit" line in light of the
following line, and fixes this issue.  This applies to the (default)
case when load_dataset(...) is called with with_times="auto" - when
with_times is given explicitly as True or False, interpretation of
blank circuit lines is straightforward.