Renames BulkCircuitList => CircuitList
The "bulk" prefix dates was initially used because we might have turned
these  objects into what are now "layouts" - objects that are used particularly
for the "bulk_" operations.  This didn't end up being the case, and a
BulkCircuitList was just a general tuple (immutable) list of circuits with
meta-data, so this commit changes the name to just CircuitList.  Note
that a CircuitList isn't actually derived from `list`, and is actually more
like a tuple of circuits.