Part one of adding circuit "occurrence ids".
This commit adds a _occurrence_id member and occurrenced_id
property to the Circuit object.  This property allows users
to distinguish otherwise identical circuits using
a hashable value.  This can be useful when equivalent circuits
need to be kept separate in dictionaries/DataSets, as the
occurrence id is used in hashing and equivalence testing.  String
representations of Circuits with occurrence ids have an additional
"@<occurrence_id>" at the end of their string (which is omitted if
the occurrence id is None, the default).