Integrates circuit occurrences into DataSets.
Removes the clunky "occurrence tag" implementation whereby a trailing
"#<number>" string label was appended to circuits to keep duplicates
separate within DataSets.  With the new occurrence id functionality
of Circuit, the "keepseparate" DataSet behavior can be neatly
implemented.  It's noteworthy that a DataSet having a non-"keepseparate"
collistion action (i.e. "overwrite" or "aggregate") will strip the
occurrence id (set it to None) of incoming circuits so that aggregation
or overwriting is performed intuitively.

This commit also fixes a few minor bugs introduced in the last
commit, having to do with occurrence support in Circuits.