Updates Circuit comparison methods (so sorting works, among other things).
The previous Circuit comparison (<, >, ==) methods would compare a
Circuit's .tup (a tuple containing line labels, etc) with tuple(right_hand_side).
But tuple(circuit) gives the Circuit's layer labels as a tuple, and is
*not* the entire circuit.tup tuple.  Comparison functions have been fixed
to properly compare circuits based on their full .tup attribute, and to
compare non-Circuits with just the Circuit's layer labels.

This fixes weird issues with sorting lists of Circuits (though the
ordering is not always what you'd expect even though it's at least
consistent now) that would occur because of the erroneous asymmetry
between circuitA < circuitB and circuitB > circuitA.