Updates `sparse` and `dense_rep` usage in cloud noise model construction.
This commit untangles the two notions of "sparseness" related to
cloudnoise (and potentially other) models.  The old `sparse` argument
has been renamed `sparse_lindblad_basis`, as it is used to indicate
whether the Hamiltonian and "other" basis used by Lindblad error generators
is sparse.  (This sparseness saves space especially for high-weight errors/gates.)
It is not stored explicitly in a model, but can be supplied to model
construction routines.

The `dense_rep` argument to a LindbladOp determines what type of
representation object underlies the operator.  If True, a dense
rep is used, which is fine when the (often embedded) Lindblad
gates act on only several qubits - so also applicable to the
case of low-weight errors/gates.  If false, a specialized Lindblad
representation is used, which is more space efficient but less
computationally efficient (at least for small sizes).  The `dense_rep`
value is now set by a `sparse_lindblad_reps` argument, which is given
to model construction and initialization routines.  Whether other types
of operations (e.g. a ComposedOp) uses a dense representation or not
is decided base on whether the current simulator is a
MatrixForwardSimulator or not.