Updates default PlaquetteGridCircuitStructure ordering.
Previously, when a PlaquetteGridCircuitStructure object was
initialized it placed all the given "additional circuits" at the
end of it's list (placing the plaquette circuits first).  When
generating GST structures, the additional circuits are LGST circuits
and we'd like to place them at the beginning of the list instead
(this is what was done prior to the recent circuit structure upgrade).
This commit adds a `additional_circuits_location` argument to
PlaquetteGridCircuitStructure.__init__, which can be "start" or "end"
and defaults to "start" -- dictating where the additional circuits are
placed.  Now the stdlists.py functions can use the "start" location to
produce the desired (and prior) LGST-circuits-first behavior.