Adds OrderedDicts into circuit structure/list classes for python 3.5.
As much of the new circuit lists/struct improvements were developed
using python 3.8, there were several instances where a dict was used
but a ordered dictionary was needed - or at least desired - so that
the ordering of generated circuit lists is deterministic.  This commit
adds OrderedDicts in place of regular dicts in several places.