Updates standard GST circuit creation so it still makes circuits with "nice" strings.
Because there was a bug/inefficiency before in Circuit.__add__ that was
fixed in this feature branch, we needed to patch the way standard GST
circuits are created to retain their "nice" strings.

Specifically, Circuit.__add__ merges string representations nicely, e.g.
Gx*4 => "(Gx)^4", only when the Circuits being added have their ._str
(cached string rep) already computed.  This isn't always the case,
as ._str is computed on demand for efficiency (previously it *was*
always computed in __add__ because __add__ used .str instead of ._str).
We don't want to go back to having __add__ always compute the string
representations, but we also want the GST circuits to have string
representation like "Gx(Gy)^4GxGx", not "Gx(GyGyGyGy)GxGx", as this
will facilitate use with CircuitLabels in the future.  So, in the
create_lsgst_circuit_lists function we explicitly call the .str
properties on all the input fiducials and germs so that we're guaranteed
to have valid ._str members in the circuits this function adds together.
This shouldn't add much overhead because the circuits used by
standard GST are typically small (1 or 2 qubit).