Fixes bug in wildcard logl objective function; adds printer to custom opt.
A bug from the flexible-fwdsim upgrade was causing the wildcard
objective functions not to work (the objective fn was using the
wrong - i.e. not-updated-by-wildcard-budget probabilities).

Secondly, the wildcard optimization (in _compute_wildcard_budget) was
failing in some cases due to SciPy's Nelder-Mead routine behaving "badly"
(meaning, non-intuitively, but in the documented way).  To fix this,
as we have in the past, the scipy.optimize call was replaced with our
custom "supersimplex" method.  This method hasn't been used in a while,
so this commit also updates our custom `minimize` routine to use a
VerbosityPrinter object and to more easily plumb arguments down to
the underlying _fmin_* optimizatin routines (all in optimize.py).

In the end, we should have a working and more robust wildcard budget
optimization implementation now.

The optimization unit tests have been updated to plumb the correct
parameters down to the "supersimplex" method - to do this a new
unit has been created.