Lots of bug fixes from running and improving system tests.
This commit fixes a number of bugs that were picked up by the system
tests (but missed by the unit tests) related to the recent framework
updates (forward-sims and circuit structures).  The system tests
were updated as well, and most of them pass now.  Notably, the
load-old-gateset test doesn't work and is skipped (in testGateSets.py)
as are some tests which require a more manual way to split a layout
(see testGateSets.py).  Some of the more important fixes included here
are:
- fixes bug that caused MatrixCOPALayouts duplicated circuits to be constructed incorrectly.
- fixes bug with MPI (split) MatrixCOPALayouts
- removes "globalIdle" from an implicit model's primitive-op-labels
- fixes issues with using Instruments - correctly expanding circuits with instruments in them.