Fixes circuit ordering bug introduced in prior commit.
When fixing a system-test-discovered bug in the circuit structures
(namely, that a structures ._additional_circuits included circuits
in its plaquettes when the additional_circuit_location = "start"),
we created another bug: only those non-plaquette circuits were added
at the beginning of the circuit list.  The current desired behavior
is that all the "additional_circuits" passed to a circuit structure
are added at the beginning of the list (when the location = "start")
even when some elements of this list are included in the plaquettes.
This is a little counterintuitive, but works well for LGST circuits
being the additional circuits.  We could change this back in the
future.  For now, this restores the previous behavior and fixes some
unit tests (and hopefully shouldn't break any of the system tests!).