Updates bootstrap dataset creation to be more efficient.
Updates to use bulk_probs instead of repeated calls to probabilities(...),
the latter which take much longer now that matrix fwdsim is split into
a "simple" and "bulk" types (the latter being the default).

This commit also updates the bootstrap system test to go to L=2^5
instead of 2^10 so it takes less time -- there doesn't seem to be
a real need to use the default 2^10 value.