Adds a few minor fixes to unit/system tests pass.
Attempts to provide a workaround for the n-qubit sequence selection
producing a different set of circuits on TravisCI's platform.  There
isn't a clear unique correct solution, but we'd like to know if the
results change, so this commit tries to test whether the produced set
of circuits is either the one generated on Erik's machines or the
one generated by Travis CI's python 3.5 environment (this is a HACK).