Fixed dict -> OrderedDict to make implicit model primitive op order deterministic.
An oversight in creating a 'globalIdle'-less dictionary (instead of
an OrderedDictionary) of primitive operations in ImplicitOpModel
caused the order given by model.primitive_op_labels to be non-deterministic
in python 3.5 and below.  This should fix the system test failure
we've been seeing - but we'll leave those debug statements for now
just in case it doesn't.