Adds a significant amount of memory-estimation framework.
Updates our memory estimation paradigm to one where we try to
collect a set of "array types" from the to-be-created objective function,
the optimizer (if present), and the forward simulator *before*
the COPA layout is created (within MDC-store creation).  This list
of array types is then converted to bytes and compared to the current
resource alloc's memory limit in a forward simulator's create_layout
method, where selecting a particular layout may be done based on the
memory limit.

Most/all of the framework exists now, but needs to be debugged.  Also,
this commit updates how "array types" are represented, using letters
to indicate dimensions, e.g. 'E', 'EP', 'e' rather than esoteric 'p',
'dp', 'hp' strings.