Adds some experimental code for making the wildcard optimization faster/more robust.
The new method attempts to monotonically walk down to a minimum by
essetially integrating a differential equation in the wildcard vector.
So far, it seems to mostly work, but doesn't run appreciably faster and
there are still some things to work out in terms of current L1 conditions
and how to deal with boundary cases.  Also, it's currently raising an
AssertionError on our test problem, so there are likely some bugs that
still need addressing.