Changes default circuit reading/writing to put parenthesis around single ops.
Previously, when a circuit was repeated (multiplied or exponentiated) by 1
this would not cause a circuit label to be formed, which is efficient but
also "special behavior" that seems better not to include.  Also, as parenthesis
are more meaningful with CircuitLabels in use, a CircuitLabel.__str__ *always*
adds parenthesis to itself so that it is read back in correctly (this was
a bug fix).  This may break some tests or other assumed behaviors, but it
seems the most sensible convention to take moving forward: whenever you
multiply a circuit (even by 1) it results in a CircuitLabel (when expansion
is set to False).