Adds guarding against nonzero derivatives at minimal objective due to bad limiting.
If v is the vector of objective function terms (e.g. delta-logl values)
then the `lsvec` function computes sqrt(v) and the derivative is
0.5/sqrt(v) * dv/dparams.  When v -> 0 the first term blows up BUT it
should aways be overcome by dv/dparams -> 0 faster, since v=0 is by
construction the exact minimum of the objective.  This limiting doesn't
always work out correctly, however, sometimes because of regularization/roundoff
so this commit adds a safeguard that explicitly sets the derivative to
zero when v < 1e-100.

This issues was causing problems when computing time-dependent objectives
in particular.