Adds .copy() methods to fast-reps and optimizations to timedep calcs.
In an effort to speed up time-dependent calculations, this commit
adds the use of a `local_repcache` holding (copies) of representations
of operators at different times.  This may get unweildy when there
are too many times, etc - so there should be a way to disable this
caching (there isn't yet).  To facilitate this, copy methods have been
added to all the DM_-type operator representation classes.  This
commit also adds support for the `outcomes_cache` added to
MapForwardSimulator in a recent commit.

Note that these additions (copy methods and outcomes_cache wiring)
have *not* been added to slowreplib.py and at least the function
signatures need to be updated (copy methods are only used for
caching which may not be used in slowreplib anyway).