Adds a "time-dependent mode" to MatrixForwardSimulator.
When MatrixForwardSimulator.create_layout is called with a DataSet
that has nontrivial time dependence, a new type of layout - a
MatrixTimeDepCOPALayout - is created, which creates a separate matrix-calc
atom for each distinct time.  Time dependent calculation methods
(e.g. bulk_fill_timedep_dchi2) are implemented for MatrixForwardSimulator,
and expect a MatrixTimeDepCOPALayout layout.  This "time dependent mode"
should probably be factored into a new ForwardSimulator object in
the future, since sometimes the usual MatrixForwardSimulator may be
desired even on time-dependent data.  At the very least, there should
be an option given to MatrixForwardSimulator.__init__ that determines
which "mode" is desired.