Upgrades "Russian peasant" method used by MatrixCOPALayout.
Adds logic to creation of a matrix layout that looks ahead to choose
not just the "bite" that adds the most to the current circuit segment
but the bite that *along with the next bite* adds the most to the
"eaten" circuit.  This look-ahead functionality causes the evaluation
tree constrution algorithm to avoid choosing a slightly better (longer)
bite at the expense of then not having a good option for the next
bite - e.g. biting off a (germ^power + part_of_fiducial) instead of
a germ^power.  In limited 1Q testing with long (8192) circuits, this
update has resulted in significantly better - i.e. more compact -
evaluation trees.

Contains commented out debug statements to be REMOVEd later.