Adds an optimized version of the "swell" function - a performance bottleneck.
The swell function is used to construct basis sets for a disturbance
calculation.  The initial implementation was straightforward but relied
on lots of numpy krons, etc.  This commit adds a new version that makes
use of pyGSTi's embedding logic to more efficiently embed a weight-k
transition matrix within a larger one, and should speed up disturbance
calculations significanty.

This commit leaves in a bunch of debug logic and print statements,
including a check within the swell function itself to compare its
output with the old version - so right now the new version isn't
any faster.  Once we're satisfied the new implementation is correct
we'll remove the debug apparatus.