Fixes bug introduced in last commit.
We should only add ('SPAM',) to the tuple of primitive ops (for a
wildcard budget) when the primitive operations haven't been specified
by the user.  In the latter case, the user may have specified a dict,
which can't be added to a tuple and will raise an error - and if the
user wants a SPAM term they should just include one manually.