Removes several assertNoWarns tests from unit tests (test_model.py)
Because a recent commit added a warning when an inefficient EvaluationTree
is created, a number of the unit tests have begun to fail.  These
tests were checking for no warnings - in particular the presence of
*other* warning about scaling issues in the matrix probability calculations.
So it's correct to remove these assertNoWarns checks, though in the
future we may want to try to check for the specific scaling warnings in
a more clever way, as these are not being checked now.