Adds a "mode" to MatrixForwardSimulator and "distribute_by_timestamp" init arg.
Adds an explicit MatrixForwardSimulator.__init__ argument,
`distribute_by_timestamp` which must be set to True (and defaults
to False) in order to engage the newly added time-dependent mode.

This commit also improves the mpitools.py sum_arrays function so
it will still work when there are more processors than timestamps
(adds the 'owners' argument), and allows a MatrixTimeDepCOPALayout
to be created even without a given time-dependent DataSet (seems like
this case would be rare, but this makes things more robust).