Fixes (*I think*) an edge case of the wildcard budget calculation.
Replaces _np.nan with 0.0 in two cases.  I think this is a correction
because when moving the alpha_break & beta_break values farther from 1.0
and one of the sets A or B runs out of elements (so sum_fA or sum_fB == 0)
then the "TVD_at_breakpt" should still be computable and should just have
no contribution from the empty A or B set.

In the previous version (with the nans), it seems that a test
case where one probability is low and the other is high causes an
issue - i.e. set A and set B each start with one element, and the
targeted TVD is small so the B element => C then the A element
get evaluated with sumB == 0.