Adds more experimental testing code for running CVXOPT and soft min/max fns.
In an ongoing effort to try to improve the wildcard optimization
(which we think we've proven is convex) this commit uses cvxopt's
intervace for convex optimization when the objective is linear and
the constraints are nonlinear (using cvxopt.cpl).  The optimizer
seems to make progress, but never actually converges.  Comparing
the supplied derivative and hessian against finite-difference
versions indicates the function, derivative, and hessian being
suppiled to cvxopt are consistent.  We thought that perhaps the
discontinuous 2nd derivative (Hessian) was the issue, and attempted
to add "soft" minimum and maximum functions but this didn't help
(i.e. cause cvxopt to converge) so they're turned off currently.
Still a WIP.